:rotating_light: Made shop item price an integer

+++ src/main/kotlin/com/github/myra/kmongo/data/guild/DbShopItem.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DbShopItem(
     val id: String,
-    val price: Long
+    val price: Int
 )
